# app/static/swagger.json
{
    "components" : {
        "securitySchemes" : {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"}
          },
...
        },
    "servers" : [ {
      "url" : "https://viccnick.azurewebsites.net"
    } ],
    "openapi" : "3.0.1",
    "paths" : {
...
      "/api/convert_certificate" : {
        "post" : {
          "description" : "Convert certificate",
          "requestBody" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "public_key" : {
                      "type" : "string"
                    },
                    "private_key" : {
                      "type" : "string"
                    },
                    "passphrase" : {
                      "type" : "string"
                    }
                  }
                },
                "examples" : {
                  "0" : {
                    "value" : "{\n    \"private_key\":\"-----BEGIN PRIVATE KEY-----\\n    MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCl4ZWR/x057L+x\\n    l9GP6QjGyPLXwIiTqj69J8EdxLZTaw3VpwbNnIlZhTwRHoC+skoQrUn6eytfxgSI\\n    qh6jXgIJ4iIr+jBFoPvl6uR1q1LFLnFm3IiGI/qIN7ew4Y0p1w9Yk9SAmUxcAqlB\\n    VTRKN4bn1Rpslv9MKouJJeM0v1qV7yEWYZO1F1QBhgTSYWJs0dgc6Z+Od2AxTktD\\n    Yiza4MGzH3m8izpcy44YHyYKMVvANMgBJttWOu/RDHFKoN2zxPcYxjucMQOfmBbC\\n    +iJV2T16KmME3E3n8srqKNr+i8kGgVoPS6I3U8TI7aehtDpf1KHD3oo3PVu5P5w2\\n    gM66v8LbAgMBAAECggEAPVKJLBk/ETBb50fEIWrZIz+bQrY65zfrDJtbMWmMq4bq\\n    bN7bzXv+bJksDPoVRyn65nJhZFdPQA8gkvMs82Dn6QZvcomBeeQa6zKsNXHZ3eoe\\n    4xNI5yn6zoumfR3MxThrRGgWCVpil6smLfBVLdCcLy0hIR6rD1m1DsTyWwgYcU+G\\n    s5XBYCqY7VybIUyO7uJvw83Y5//HHBza59jkiyuUiCLwIQhk/ESg+B34got9sfju\\n    9ydoleUf8arXbvfXNuN9/qP7h5HXVVmCnvUoR3d68FAP1QlmOBtU7PvB1sKd/p2s\\n    4ULtClr7PMSD19kFjNC2gP89Z3vAwsIfZvUwCR9YlQKBgQDN4O3hgAvpPU81bEtg\\n    A+T1RkNj6GGG9TE7w7Bu4DIRptTWTLI6XyAgZHpR3sKil+2ueRU2k5zpO8usfAvA\\n    ywj6Lp7vCcIP0Jrcnw5gXlZoDcE7PKLQAbjGWq1p/61zCX1BV7OrTKQawd3z+3CJ\\n    0VTFPaLShvyX6356xOBZkdhELwKBgQDOQ+AsGm2IBooqA0mDbXBIxrrPHdVm/a0p\\n    OzxtuzKPkc9razwxXdAlBdCezaj5T/AQaAsNh3luKdNbWZevFslIJu1QSRvHdtWf\\n    yP+da9h12gRhkGODF7rsbbRVVRCyzVPbnJqAqKJ99m6paXdW/yVss7w/4hKIqFhN\\n    znU23JbFFQKBgDzHrvLn7pxJMV1AP3B1Nd8Dg1m3fTNup75RI4FYRu/XJPavZd69\\n    vqpRf5sz/Jor6EueC5k1MxnYQWUGqiGcz2Ws5dcENnqcx1tiycfBkhjeFrH6UMa/\\n    +2Dz3EZy6S4RGx1zWuOzmK8dlP02uh614iTo8Wzd1Sy0pX5xKRrIZ+SZAoGAIWFe\\n    sAR+4VqdlXr+tKRLXZu7s7USSh6XnWqWtvNqlU11kd7C2QJjovoz7t89j0tegGk2\\n    ELRx9GAUcl4nbBB19pVGJo4OTtNucvNmWqWPXFMyDk/9qsN2XrjKTFz93NHZ1lqb\\n    rmKcGWJ+OjB1PcEA33ye5H/DnBYSZFUbJwu/4pkCgYEAoz1Y2YpB64qN57qGxPxX\\n    HAjQui7ALbAaBZ6SF4LwTnXxJ2K+AG/xdhSYiV11BnwNpSx/DHxi8CM44pnOoLJT\\n    yhl5MAmUGfLB4S8Oh4XycIdFKcezXKGMphNP6tXHZ43CFPtQf3AW89Uy9k0ZzFS7\\n    8jmDUVABZ9ZgC7hObz4pnRk=\\n-----END PRIVATE KEY-----\\n    \",\n    \"public_key\":\"-----BEGIN CERTIFICATE-----\\n    MIIDjTCCAnWgAwIBAgICCLswDQYJKoZIhvcNAQELBQAwWjELMAkGA1UEBhMCVVMx\\n    EzARBgNVBAgTCldhc2hpbmd0b24xEDAOBgNVBAcTB1NlYXR0bGUxJDAiBgNVBAoT\\n    G2dldGFDZXJ0IC0gd3d3LmdldGFjZXJ0LmNvbTAeFw0yMzAzMjYwOTExMjZaFw0y\\n    MzA1MjUwOTExMjZaMGsxCzAJBgNVBAYTAkNIMQ0wCwYDVQQIDARCZXJuMQ0wCwYD\\n    VQQHDARCZXJuMRAwDgYDVQQKDAdUZXN0T3JnMSwwKgYDVQQDDCNua2Rhc25kYXNk\\n    c2FkYWRhc2FzZGFzZHNhZC5wbmV0LmNobjCCASIwDQYJKoZIhvcNAQEBBQADggEP\\n    ADCCAQoCggEBAKXhlZH/HTnsv7GX0Y/pCMbI8tfAiJOqPr0nwR3EtlNrDdWnBs2c\\n    iVmFPBEegL6yShCtSfp7K1/GBIiqHqNeAgniIiv6MEWg++Xq5HWrUsUucWbciIYj\\n    +og3t7DhjSnXD1iT1ICZTFwCqUFVNEo3hufVGmyW/0wqi4kl4zS/WpXvIRZhk7UX\\n    VAGGBNJhYmzR2Bzpn453YDFOS0NiLNrgwbMfebyLOlzLjhgfJgoxW8A0yAEm21Y6\\n    79EMcUqg3bPE9xjGO5wxA5+YFsL6IlXZPXoqYwTcTefyyuoo2v6LyQaBWg9LojdT\\nxMjtp6G0Ol/UocPeijc9W7k/nDaAzrq/wtsCAwEAAaNMMEowCQYDVR0TBAIwADAR\\n    BglghkgBhvhCAQEEBAMCBPAwCwYDVR0PBAQDAgWgMB0GA1UdJQQWMBQGCCsGAQUF\\n    BwMCBggrBgEFBQcDATANBgkqhkiG9w0BAQsFAAOCAQEANLRxBqjbnbOh6m0ijqOG\\n    wYqbwEDQ+xasnAWWCxNPIRYqHGsU2b1jpWqj+IyWhP7VziprxtiGB0QRT4Ejkrsc\\n    drYG9WHeknny9rUTuiUaKIMbi5wH4EVVDVruaSJaRY7X/k7J3WbykXvJoCcuJbVg\\n1Nl69Nbc+8SBFnxIs/O/SQnfEITNKdHvOUQXhVSLntEufn+H0OGNZbBlTmZHZAl1\\n    /TdOVftdGq3VV2FNNDBtLbb4qc30KoqOdx6JFWwjSK7qSeTDtZAuP7O7ZrgpNGkH\\n    RtyVwk5NDRGKr0VLzMH6kSPmsqShRWM1tx1wRgbuuXHioSmviDSULRIs8gQxeN2d\\n+w==\\n-----END CERTIFICATE-----\",\n    \"passphrase\": \"testpassword\"\n}"
                  }
                }
              }
            }
          },
          "responses" : {
            "200" : {
              "description" : "success",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "data" : {
                        "type" : "object",
                        "properties" : {
                          "pfx" : {
                            "type" : "string"
                          }
                        }
                      },
                      "status" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"data\":{\"pfx\":\"MIIJ3QIBAzCCCZcGCSqGSIb3DQEHAaCCCYgEggmEMIIJgDCCBDcGCSqGSIb3DQEHBqCCBCgwggQkAgEAMIIEHQYJKoZIhvcNAQcBMBwGCiqGSIb3DQEMAQMwDgQIDCJyHm6pDAsCAggAgIID8JH7RGK011ZLRkD8uqubuql+PNZsla2unjyYcCu/KU+y7XqUwM4oOqvvqB64/BVa3GZqkjq4/em+8sa6uGUyxbzwIwg2tx7pq1zNTXAYRc5F6Jr9V70jsXtoskY6vT8oIa7gsku02PEY13HpThW92DOoZu2YhRnxrbMWbERonRyMkyUcloViBlImh8806qq/krBJYvjC36ZW9Q96A0W+gmDIWIVTb7ixz0I5IBgKrR08D9dZ6NFQ4KYR0RuQKv2aelvkak3aJ6IN2ALPDeTt3gWXchb6cYOasSpKiEPvoWxqUHOXRsqcVJOfR6H0YGwHs7hCrxZqeKDJN6knP0POfPotM/z2aQKEGe2XHF0A52b2j+OA89iOhPGs0Quz7pClb8scMt/D/cmy7eKAecKr9WPoKxtUYh9liaZYLf1XXv9RB0TQOmhZ7MGZ3+lvvHUigoLz5SRYwryNPAEp5DHE0bM+THeA9SYcs9leQdydu3Wd0umqVeJiW/tWVEnQtuR+waPZmE//IEJeCOLIBzekvdgu0E4eHW5MtJRi1bvKFHG4VmxEEFRur5aSP6ksxl85pcIl1jd3+ytyic3tn46kfoCVrUPnrKDfq574Gr/Knr2PFFFl5L3MrC8kSCNjevEyRrVgrwVr6cM/FHGlidMdDojFlduUrexxZRyfGMWkjWbPvzkORiJugt1FfLIT34ES0y+dHIb3qHEN3zHBOQpdnenxfl2s1fZU9o3t6cf2Qa8EJYGaoqsBY84tEKOFOm2nCb0CqdlkJuXtwvHKyOr0hvnkKG7JxfaTJgmnwa6VopB2st4cd2HP2gUFc3djz27xjNJwKGCRvJvcmC2/exJjW8wuZEXxdktc3hVCa3Kx8wWdtlB343i7As2CzN5k5cUSZ338VjxnoECN/4Zo4EPTZsD4IfP7Ie92cclT3aeuSzOgzXjY6N3xf9k55F4Ch7S76S82BoBOOtNT0DCGtn+L3LI89mfdiC5Yr8gzqzk+FlPQiWEPtE/G3H9eD0r2FNnio1BtVeYHjeVcEUl+45gCxX2JJHOkFAWgT8dCv+Obo1Ru9ZI23aiqpKieGErSMHg/6xjTp8XIUGShX1JuEoFHtmVPtf6m5WQmHaQs2065idAnSenOG+yz3Mici1VjtrecUxVzX5jcCpnbXN3/eEwUyhDi7KJ+wRWpLd7zXCZwOVYi2ZnwKxSFEDTQnXBuMU/Hi4lj1Vks7q3lkh0gvHcJnWAhSLGA28Ej/zWojp+OrC18gFP2f3os0zrNAHUM/Ze8ntFtnqM+YDQdEhZO6ej6W2RH3cMDGdJn13YF+zk3mnMRN+EGG2VB6lwTe+eC0V+DHzCCBUEGCSqGSIb3DQEHAaCCBTIEggUuMIIFKjCCBSYGCyqGSIb3DQEMCgECoIIE7jCCBOowHAYKKoZIhvcNAQwBAzAOBAgkI0521FA9gwICCAAEggTIlFk/KtUIIJd2eEtezBlZK67UEgAUvZDqbu6wsoAi3v5DsSLkppGZEJaEBUenlIQ9kKSAj1JwkPg+rvr1iQqJSUeA8v56LpPnFNZCSdasE8sh6jQorPjAu95YGbgehmqqgqYXTqPPrRS6a2aVr6SXY/bE+vyanJamyNun6Bl/QJeC1tEBu4vNohsWqukIhWgwPZ7X7GlhOaiO8U3KLg+CZL99r/kCND2WWa46uY/uxMq+JHZgBzzG6pZq1pikNsngKwO8oeaP1FTUmKUqODbZqSPn9VeJ7tKBZsqcuc1Uu1vnNqtNEMR5I+7DbynJ9/+gDfXq8S4pTaU7LEv+0GKJ6wHgFKFvwJpdu+IVc/O7aqmjeQILVDAttYVVGJbCBh/QI+m4jQ1ZNN0nduAjwrZ/u1bfZxT07emHBnpfVk1tM2lRkgeTfKe5v9zsWvlugNqoRJ4H0PQ7yaL0BkiPQzXIGtbzp+UZl/7QodctWo1LoykoOZrfxxRnxibW9g4AzWOvI27JtrnzesFqcSJt/FDaoOkvVFGGtokGDMK//ZkNx9gGoWDEW81GV18gG0mIOULeBem12rMqRpp6QXmKF45SnqPOYQ9HG7r59pJumcWxfnJU84q2Ql3+iasASCwbSJq8UVytZa/EHbnf+1sZMqyrZtZ24rglmkPOJ8sOk+PX9J1NVfOf7JHGniAObydHJg5nfkZNidnc0ux8h5W2o3BkbYeO9RcT3gbAciiwi5fXOLvfGeuYgMmhId14vV3SZ44H5x6GdjbCd/8HyFW9/gsRkeBpYcjsxtPwMUz3w/xIsOunTstgBulp+/QEJCKgR+v/lVU5ICChnc6hD/eJj+TM9bTlv/Fhp4AjxpuxhwslpQKagh8Ch+bJdn7J0C2BV4qMz7whO5RVCX9IruDCRaR9P3NlBqcEa/nBKvcv4Lx7IUJcPk9jMbUw53E/VISYr0xyQaRaERBnjyiErZfqpIbmlTG3/Dvd1JmFUsJiWTHYeyvc+XamVYPUlFcoaTA1N2dpnWgaIOGu+Xt6AHGIQy+vIuTHjNdayC28c6LO/neTdG5UqbaPM3Hu7xL8SvtK4SWUHOAWRsxjr2GwVx04LVtJc0b2bD2Tj1v0smxAqARI/Nd8/ikbyxPGB2u+mr/rspcl2d4K+8IBcKx1OhFB/310wzLdN/XxmPjfGdYFAmijwm/EyGd1QTqS/eZpb+dm7rOQ8KoUwwVv09eF9M954Ygeu/T5shBlX/MfwqbGHbNysA7mX2QCQ2C61Ryn5Funx6ltWsMG2//5eQv0paPPUX4dn4ej9s5Hzo2I0V12ltLufkgdhglAfP5jQ5RGHRR0re1xLPnWAONgBsGOKMzq1UPPVuTlfq1ThlgwNlaXE646jMJyzwy2FXqGmC4pnsVxHQ/Xg0xECyUTlVTeRKfIBXa6HJ4HhKupO9vn5ZKMSp8o3sxjsP7wvZetfz0jxQoyLkXaJiVgVMqTO2IAAn+2TWoT18wCqIHN/jIyuD9qoaUM1OUwhzFEmivHEiVHCm1uNFVz/X/s1NKBK4iQZ4M6eFC9S5S9whACdXvWOU9x3eBhsnNFr9oWiSA9NR1GQOfW6kW24jUbIWAXAkdyHy9G6S6/INZ0nFjEZYjVMSUwIwYJKoZIhvcNAQkVMRYEFEW/U5dtYo+BXGHwukZIbx81jZsBMD0wMTANBglghkgBZQMEAgEFAAQgPG+GSMHahdPDtmG8okqidEVHb64G8DOJeLF9onO3JBoECGknwtJvd17z\"},\"status\":\"success\"}\n"
                    }
                  }
                }
              }
            },
            "400" : {
                "description" : "Bad Request",
                "content" : {
                  "application/json" : {
                    "schema" : {
                      "type" : "object",
                      "properties" : {
                        "error" : {
                          "type" : "string"
                        }
                      }
                    },
                    "examples" : {
                      "0" : {
                        "value" : "{\"error\":\"Bad Request\"}\n"
                      }
                    }
                  }
                }
            },
            "500" : {
                "description" : "Internal Server Error",
                "content" : {
                  "application/json" : {
                    "schema" : {
                      "type" : "object",
                      "properties" : {
                        "error" : {
                          "type" : "string"
                        }
                      }
                    },
                    "examples" : {
                      "0" : {
                        "value" : "{\"error\":\"Internal Server Error\"}\n"
                      }
                    }
                  }
                }
            }
          }
        }
      },
...
      }
    }